Added alternate rear image for empty rear images after creating cards.
You can now have a front and rear card side that can be filled with text
and images. If the rear side is not changed through new text, the user
can now add an alternate card image, that will replace the empty rear
image for the card.
The alternate rear image is applied in a post process step.