Added inline bold (<b></b>) and inline italic(<i></i>) font styles
Added the possibility to mark text passages bold or italic.  This
includes a massive rework of the draw functions of the class Field.
Fixed the Field Mapping tool to not end with an ClassCastException by
casting the initial values to int.
You can now click on either of the preview images to add and edit the
fields.